fix: add file checker at task 4

+++ 04-copy-directory/index.js
@@ -2,12 +2,12 @@ const { mkdir, copyFile, readdir } = require('fs/promises');
 const path = require('path');
 
 const targetFolder = path.join(__dirname, 'files');
-const folderPath = path.join(__dirname, 'dir');
+const folderPath = path.join(__dirname, 'files-copy');
 
 async function makeDir() {
   try {
-    const folder = await mkdir(folderPath, { recursive: true });
-    console.log(folder);
+    await mkdir(folderPath, { recursive: true });
+
     copyToDir();
   } catch (err) {
     console.error(err.message);
@@ -18,11 +18,12 @@ async function copyToDir() {
     const files = await readdir(targetFolder, { withFileTypes: true });
 
     files.forEach((file) => {
-      console.log(file);
-      copyFile(
-        path.join(targetFolder, file.name),
-        path.join(folderPath, file.name)
-      );
+      if (file.isFile()) {
+        copyFile(
+          path.join(targetFolder, file.name),
+          path.join(folderPath, file.name)
+        );
+      }
     });
   } catch (err) {
     console.error(err.message);
